LayoutTests/imported/w3c/web-platform-tests/css/css-tables/visibility-collapse-rowcol-001-expected.txt

Visibility collapse

Spec
Setting row and column to visibility:collapse changes both height and width. The collapsed result should have whitespace in the bottom right corner.


FAIL spanning col visibility:collapse changes table width assert_equals: spanning col visibility:collapse changes table width expected 112 but got 148
FAIL spanning row visibility:collapse changes height assert_equals: spanning row visibility:collapse changes height expected 112 but got 148
